requirements.txt
streamlit-navigation-bar
streamlit-community-navigation-bar
plotly.express
pandas
requests
os